feat: connect the editScreen to the backend
The profile edit screen now uses the ProfilesViewModel to get the user
from the back-end, and update it when on Firebase when it is updated
on the phone.
The asynchronous loading of the user is partially implemented. The
TagsViewModel has complicated things a little, and during next sprint,
I'll create a loading screen that will load everything before
displaying the page.