fix(PP): revert 6ca9b65 (uploadCurrentUserProfilePicture signature change)
Previously I changed the signature of the uploadCurrentUserProfilePicture method so I
could match the edition style Louis used in ProfileEdit.kt. However, I realized that
it was a bad idea to put the image conversion logic in the UI layer, so I moved it back
to the viewmodel. This means that discarding the change while editing profile does not
discard the change of profile picture.